+++ kimchi/src/circuits/berkeley_columns.rs
@@ -1,7 +1,10 @@
-use crate::circuits::{
-    expr::{self, Domain, GenericColumn},
-    gate::{CurrOrNext, GateType},
-    lookup::lookups::LookupPattern,
+use crate::{
+    circuits::{
+        expr::{self, ColumnEvaluations, Domain, ExprError, GenericColumn},
+        gate::{CurrOrNext, GateType},
+        lookup::lookups::LookupPattern,
+    },
+    proof::{PointEvaluations, ProofEvaluations},
 };
 use serde::{Deserialize, Serialize};
 use CurrOrNext::{Curr, Next};
@@ -88,6 +91,29 @@ impl expr::Variable<Column> {
         match self.row {
             Curr => format!("Curr({col})"),
             Next => format!("Next({col})"),
+        }
+    }
+}
+
+impl<F: Copy> ColumnEvaluations<F> for ProofEvaluations<PointEvaluations<F>> {
+    type Column = Column;
+    fn evaluate(&self, col: Self::Column) -> Result<PointEvaluations<F>, ExprError<Self::Column>> {
+        use Column::*;
+        let l = self.lookup.as_ref().ok_or(ExprError::LookupShouldNotBeUsed);
+        match col {
+            Witness(i) => Ok(self.w[i]),
+            Z => Ok(self.z),
+            LookupSorted(i) => l.map(|l| l.sorted[i]),
+            LookupAggreg => l.map(|l| l.aggreg),
+            LookupTable => l.map(|l| l.table),
+            LookupRuntimeTable => l.and_then(|l| l.runtime.ok_or(ExprError::MissingRuntime)),
+            Index(GateType::Poseidon) => Ok(self.poseidon_selector),
+            Index(GateType::Generic) => Ok(self.generic_selector),
+            Permutation(i) => Ok(self.s[i]),
+            Coefficient(i) => Ok(self.coefficients[i]),
+            LookupKindIndex(_) | LookupRuntimeSelector | Index(_) => {
+                Err(ExprError::MissingIndexEvaluation(col))
+            }
         }
     }
 }
